+++ src/config/src/tantivy/query/mod.rs
@@ -13,5 +13,5 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-pub mod collector;
 pub mod contains_query;
+pub mod ids_collector;
